requirements.txt
pytest==7.1.3
pytest-mock==3.11.1
pytest-cov==4.1.0
pylint==2.17.5
flake8==6.1.0
isort==5.12.0
mypy==1.5.1
colorama==0.4.6
customtkinter==5.2.0
requests==2.31.0
tk==0.1.0
pyyaml==6.0.1
Pillow==10.0.0
bcrypt==4.0.1
pandas==2.1.1
